requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.6.2
cffi==1.16.0
channels==4.1.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cryptography==42.0.8
daphne==4.1.2
distro==1.9.0
dj-database-url==2.1.0
Django==4.2.7
django-appconf==1.0.6
django-cors-headers==4.4.0
django-crispy-forms==2.2
django-widget-tweaks==1.5.0
django_compressor==4.5
djangorestframework==3.15.2
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.66.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
openai==1.54.4
oscrypto==1.3.0
packaging==25.0
Pillow==10.4.0
proto-plus==1.25.0
protobuf==5.29.0
psycopg[binary]>=3.0,<4.0  # use psycopg3 binary driver compatible con Python 3.13
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyHanko==0.25.0
pyhanko-certvalidator==0.26.3
PyJWT==2.9.0
pyOpenSSL==24.2.1
pyparsing==3.2.0
pypdf==4.2.0
pypng==0.20220715.0
python-dotenv==1.0.0
PyYAML==6.0.1
qrcode==7.4.2
rcssmin==1.1.2
reportlab==4.2.0
requests==2.31.0
rjsmin==1.2.2
rsa==4.9
service-identity==24.1.0
setuptools==75.1.0
sniffio==1.3.1
sqlparse==0.5.0
tqdm==4.66.5
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.2.2
uvicorn==0.35.0
websockets==13.1
whitenoise==6.6.0
zope.interface==7.1.0
